Add 4th Job Skill Related Items / Added Item Requirements To Working Skills
=Database
*Added all 4th job skill related items to the item_db.

*Added Meister craftable items to the produce_db.

=Skills
*MT_M_MACHINE
-Added support for this skill.

*MT_A_MACHINE
*MT_D_MACHINE
*WH_HAWK_M
-Added item requirements.

+++ src/tool/csv2yaml.cpp
@@ -707,7 +707,7 @@ static bool parse_item_constants_txt( const char* path ){
 		str[21] = str[21] + 1;  //skip the first left curly
 		*p = '\0';              //null the last right curly
 
-		uint16 item_id = atoi( str[0] );
+		uint32 item_id = atoi( str[0] );
 		char* name = trim( str[1] );
 
 		aegis_itemnames[item_id] = std::string(name);
